combined users into app. updated tests to include permissions check.
