+++ app/tests.py
@@ -1,5 +1,7 @@
 from django.test import TestCase
 from django.contrib.auth import get_user_model
+from rest_framework.authtoken.models import Token
+from rest_framework import status
 from .models import Project, Node, UserMembership, NodeMembership
 
 
@@ -9,6 +11,12 @@ User = get_user_model()
 class TestApp(TestCase):
 
     def testListAccess(self):
+        user = User.objects.create(username="admin", is_staff=True)
+        admin_token = Token.objects.create(user=user)
+
+        user = User.objects.create(username="user")
+        user_token = Token.objects.create(user=user)
+
         self.setUpMembershipData(
             profile_membership=[
                 ("ada", "sage", {"can_develop": True}),
@@ -30,24 +38,33 @@ class TestApp(TestCase):
             ],
         )
 
+        # require auth for this endpoint
         r = self.client.get("/profiles/ada/access")
-        self.assertEqual(r.status_code, 200)
+        self.assertEqual(r.status_code, status.HTTP_401_UNAUTHORIZED)
+        
+        # require admin permissions for this endpoint
+        r = self.client.get("/profiles/ada/access", HTTP_AUTHORIZATION=f"Sage {user_token}")
+        self.assertEqual(r.status_code, status.HTTP_403_FORBIDDEN)
+
+        # check responses
+        r = self.client.get("/profiles/ada/access", HTTP_AUTHORIZATION=f"Sage {admin_token}")
+        self.assertEqual(r.status_code, status.HTTP_200_OK)
         self.assertEqual(r.json(), [
             {"vsn": "W001", "access": ["develop", "schedule"]},
             {"vsn": "W002", "access": ["develop"]},
             {"vsn": "W003", "access": ["develop"]},
         ])
 
-        r = self.client.get("/profiles/jed/access")
-        self.assertEqual(r.status_code, 200)
+        r = self.client.get("/profiles/jed/access", HTTP_AUTHORIZATION=f"Sage {admin_token}")
+        self.assertEqual(r.status_code, status.HTTP_200_OK)
         self.assertEqual(r.json(), [
             {"vsn": "W001", "access": ["schedule"]},
             {"vsn": "W002", "access": ["develop"]},
             {"vsn": "W003", "access": ["develop", "schedule"]},
         ])
 
-        r = self.client.get("/profiles/tom/access")
-        self.assertEqual(r.status_code, 200)
+        r = self.client.get("/profiles/tom/access", HTTP_AUTHORIZATION=f"Sage {admin_token}")
+        self.assertEqual(r.status_code, status.HTTP_200_OK)
         self.assertEqual(r.json(), [
             {"vsn": "W001", "access": ["access_files", "develop", "schedule"]},
             {"vsn": "W002", "access": ["develop"]},
@@ -55,8 +72,11 @@ class TestApp(TestCase):
         ])
     
     def testAccessNotExist(self):
-        r = self.client.get("/profiles/nothere/access")
-        self.assertEqual(r.status_code, 404)
+        user = User.objects.create(username="admin", is_staff=True)
+        token = Token.objects.create(user=user)
+
+        r = self.client.get("/profiles/nothere/access", HTTP_AUTHORIZATION=f"Sage {token}")
+        self.assertEqual(r.status_code, status.HTTP_404_NOT_FOUND)
 
     def setUpMembershipData(self, profile_membership, node_membership):
         for username, projectname, access in profile_membership:
