+++ config/settings/base.py
@@ -12,11 +12,10 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "rest_framework",
     "rest_framework.authtoken",
-    "users",
     "app",
 ]
 
-AUTH_USER_MODEL = "users.User"
+AUTH_USER_MODEL = "app.User"
 
 MIDDLEWARE = [
     "django.middleware.security.SecurityMiddleware",
