+++ config/urls.py
@@ -21,6 +21,5 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("api-auth/", include("rest_framework.urls")),
     path("", include("app.urls")),
-    path("", include("users.urls")),
     path("", TemplateView.as_view(template_name="index.html")),
 ]
