--- users/admin.py
@@ -1,17 +1,0 @@
-from django.contrib import admin
-from django.contrib.auth import get_user_model
-from django.contrib.auth import admin as auth_admin
-
-User = get_user_model()
-
-
-@admin.register(User)
-class UserAdmin(auth_admin.UserAdmin):
-    fieldsets = (
-        (None, {"fields": ("username", "password")}),
-        ("Personal info", {"fields": ("name", "email", "organization", "bio", "ssh_public_keys")}),
-        ("Permissions", {"fields": ("is_active", "is_staff", "is_superuser", "groups", "user_permissions")}),
-        ("Important dates", {"fields": ("last_login", "date_joined")}),
-    )
-    list_display = ("username", "name", "is_superuser")
-    search_fields = ("username", "name")
